requirements.txt
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
binance==0.3
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
client==0.0.1
constantly==15.1.0
cryptography==2.8
dateparser==0.7.2
Flask==1.1.1
Flask-WTF==0.14.2
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10.3
loginform==1.2.0
lxml==4.4.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.18
numpy==1.17.4
pandas==0.25.3
protobuf==3.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
python-binance==0.7.4
python-dateutil==2.8.1
pytz==2019.3
regex==2019.11.1
requests==2.22.0
service-identity==18.1.0
six==1.13.0
Twisted==19.10.0
txaio==18.8.1
tzlocal==2.0.0
urllib3==1.25.7
Werkzeug==0.16.0
WTForms==2.2.1
zope.interface==4.7.1
